requirements.txt
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
alembic==1.11.1
altair @ file:///tmp/build/80754af9/altair_1599835197802/work
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.4
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1619597386817/work
certifi @ file:///croot/certifi_1683875369620/work/certifi
certipy==0.1.3
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646038465422/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
comm==0.1.3
conda==4.12.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1649087926789/work
cryptography @ file:///tmp/build/80754af9/cryptography_1639400846433/work
debugpy==1.6.7
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
docker==6.1.2
dockerspawner==12.1.0
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
escapism==1.0.1
executing==1.2.0
fastjsonschema==2.16.3
fqdn==1.5.1
frozenlist==1.3.3
gitdb @ file:///tmp/build/80754af9/gitdb_1617117951232/work
GitPython @ file:///croot/gitpython_1674662789231/work
greenlet==2.0.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets @ file:///croot/ipywidgets_1679394798311/work
isoduration==20.11.0
jedi==0.18.2
Jinja2 @ file:///croot/jinja2_1666908132255/work
json5==0.9.14
jsonpointer==2.3
jsonschema @ file:///croot/jsonschema_1676558650973/work
jupyter-events==0.6.3
jupyter-telemetry==0.1.0
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core @ file:///croot/jupyter_core_1679906564508/work
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterhub @ file:///bitnami/blacksmith-sandox/jupyterhub-3.1.1
jupyterhub-firstuseauthenticator==1.0.0
jupyterhub-hmacauthenticator==1.0
jupyterhub-idle-culler==1.2.1
jupyterhub-kubespawner==5.0.0
jupyterhub-ldapauthenticator==1.3.2
jupyterhub-ltiauthenticator==1.5.0
jupyterhub-nativeauthenticator==1.1.0
jupyterhub-systemdspawner==0.17.0
jupyterhub-tmpauthenticator==0.6
jupyterlab @ file:///bitnami/blacksmith-sandox/jupyterlab-3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets @ file:///croot/jupyterlab_widgets_1679055282532/work
jupyterlab_server==2.22.1
kubernetes-asyncio==24.2.3
ldap3==2.9.1
Mako==1.2.4
markdown-it-py @ file:///croot/markdown-it-py_1684279902645/work
MarkupSafe==2.1.2
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl @ file:///opt/conda/conda-bld/mdurl_1659716024347/work
mistune==2.0.5
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multidict==6.0.4
mwoauth==0.3.8
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
notebook @ file:///bitnami/blacksmith-sandox/jupyter-notebook-6.5.4
notebook_shim==0.2.3
nullauthenticator==1.0.0
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1682520569166/work
oauthenticator==15.1.0
oauthlib==3.2.2
onetimepass==1.0.1
packaging==23.1
pamela==1.0.0
pandas==1.5.3
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pkgutil_resolve_name @ file:///opt/conda/conda-bld/pkgutil-resolve-name_1661463321198/work
platformdirs==3.5.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==11.0.0
pyasn1==0.5.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycurl==7.45.2
pydeck @ file:///tmp/abs_c06301fd-7800-4748-8e9e-0f9ab34bc0cess8ywq6h/croots/recipe/pydeck_1657894564709/work
Pygments @ file:///croot/pygments_1684279966437/work
PyJWT==2.7.0
Pympler @ file:///tmp/build/80754af9/pympler_1602785470644/work
PyMySQL==1.0.3
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3
PyYAML @ file:///croot/pyyaml_1670514731622/work
pyzmq @ file:///croot/pyzmq_1682697643292/work
requests==2.30.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich @ file:///croot/rich_1684282154404/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
semver @ file:///tmp/build/80754af9/semver_1603822362442/work
Send2Trash==1.8.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.14
sqlalchemy-cockroachdb==2.0.1
stack-data==0.6.2
statsd==4.0.1
streamlit @ file:///croot/streamlit_1673469525582/work
sudospawner==0.5.2
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli==2.0.1
toolz @ file:///croot/toolz_1667464077321/work
tornado==6.3.2
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets==5.9.0
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
tzlocal @ file:///croot/tzlocal_1668627599277/work
uri-template==1.2.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
validators @ file:///tmp/build/80754af9/validators_1612286467315/work
watchdog @ file:///tmp/build/80754af9/watchdog_1638366565112/work
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension @ file:///croot/widgetsnbextension_1679313860248/work
y-py==0.5.9
yarl==1.9.2
ypy-websocket==0.8.2
zipp==3.15.0
